Refactor: prompt separation without service
