+++ src/main/java/turing/turing/domain/gpt/GptService.java
@@ -1,4 +1,4 @@
-package turing.turing.domain.gpt.service;
+package turing.turing.domain.gpt;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
